Add the first arrangement - Ratings Summary
The arrangement pattern is still a work in progress, but this
is a very strong start at having something.
